refactor(audit-ledger): move audit ledger logic to library crate
Move core audit ledger implementation from `nono-cli` to the `nono` library
crate. This extracts shared ledger logic, making it reusable by other
consumers of the `nono` library and providing a clearer separation of
concerns.

Key components moved include:
- `LedgerRecord`, `LedgerVerificationResult`, and related structs
- `compute_session_digest` and session payload serialization
- `validate_ledger_session_id`
- `append_session_to_ledger_file` and `verify_session_in_ledger_reader`
- Ledger chain hashing logic (`hash_ledger_link`)

The `nono-cli` now uses the newly exposed public API from `nono::audit`
to manage the audit ledger. New domain separators were introduced in `nono`
for session digests and ledger chain links, and dedicated unit tests for
the ledger API were added to the `nono` crate.

Resolves:  #1114
