+++ crates/nono-cli/src/audit_ledger.rs
@@ -1,118 +1,18 @@
 use crate::audit_session::audit_root;
 use nix::fcntl::{Flock, FlockArg};
-use nono::undo::{
-    AuditAttestationSummary, AuditIntegritySummary, ContentHash, NetworkAuditEvent, SessionMetadata,
+use nono::audit::{
+    LedgerRecord, LedgerVerificationResult, append_session_to_ledger_file,
+    missing_ledger_verification_result, validate_ledger_session_id,
+    verify_session_in_ledger_reader,
 };
+use nono::undo::SessionMetadata;
 use nono::{NonoError, Result};
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
 use std::fs::OpenOptions;
-use std::io::{BufRead, BufReader, Read, Seek, SeekFrom, Write};
-#[cfg(unix)]
-use std::os::unix::ffi::OsStrExt;
+use std::io::BufReader;
 use std::path::PathBuf;
 
 const AUDIT_LEDGER_FILENAME: &str = "ledger.ndjson";
 const AUDIT_LEDGER_LOCK_FILENAME: &str = "ledger.lock";
-const SESSION_DIGEST_DOMAIN: &[u8] = b"nono.audit.session-digest.alpha\n";
-const LEDGER_CHAIN_DOMAIN: &[u8] = b"nono.audit.ledger.chain.alpha\n";
-const LEDGER_HASH_ALGORITHM: &str = "sha256";
-
-#[derive(Serialize)]
-struct SessionDigestPayload<'a> {
-    session_id: &'a str,
-    started: &'a str,
-    ended: &'a Option<String>,
-    command: &'a [String],
-    executable_identity: Option<ExecutableIdentityDigestPayload>,
-    tracked_paths: Vec<Vec<u8>>,
-    snapshot_count: u32,
-    exit_code: &'a Option<i32>,
-    merkle_roots: &'a [ContentHash],
-    network_events: &'a [NetworkAuditEvent],
-    audit_event_count: u64,
-    audit_integrity: &'a Option<AuditIntegritySummary>,
-    audit_attestation: &'a Option<AuditAttestationSummary>,
-}
-
-#[derive(Serialize)]
-struct ExecutableIdentityDigestPayload {
-    resolved_path: Vec<u8>,
-    sha256: ContentHash,
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-pub(crate) struct LedgerRecord {
-    pub(crate) sequence: u64,
-    pub(crate) prev_chain: Option<ContentHash>,
-    pub(crate) session_id: String,
-    pub(crate) session_digest: ContentHash,
-    pub(crate) completed_at: String,
-    pub(crate) chain_hash: ContentHash,
-}
-
-#[derive(Serialize)]
-struct LedgerLinkPayload<'a> {
-    sequence: u64,
-    session_id: &'a str,
-    session_digest: ContentHash,
-    completed_at: &'a str,
-}
-
-#[derive(Serialize)]
-pub(crate) struct LedgerVerificationResult {
-    pub(crate) hash_algorithm: String,
-    pub(crate) entry_count: u64,
-    pub(crate) session_digest: ContentHash,
-    pub(crate) session_found: bool,
-    pub(crate) session_digest_matches: bool,
-    pub(crate) ledger_chain_verified: bool,
-    pub(crate) ledger_head: Option<ContentHash>,
-}
-
-pub(crate) fn compute_session_digest(metadata: &SessionMetadata) -> Result<ContentHash> {
-    let payload = SessionDigestPayload {
-        session_id: &metadata.session_id,
-        started: &metadata.started,
-        ended: &metadata.ended,
-        command: &metadata.command,
-        executable_identity: metadata.executable_identity.as_ref().map(|identity| {
-            ExecutableIdentityDigestPayload {
-                resolved_path: path_bytes(&identity.resolved_path),
-                sha256: identity.sha256,
-            }
-        }),
-        tracked_paths: metadata
-            .tracked_paths
-            .iter()
-            .map(|path| path_bytes(path))
-            .collect(),
-        snapshot_count: metadata.snapshot_count,
-        exit_code: &metadata.exit_code,
-        merkle_roots: &metadata.merkle_roots,
-        network_events: &metadata.network_events,
-        audit_event_count: metadata.audit_event_count,
-        audit_integrity: &metadata.audit_integrity,
-        audit_attestation: &metadata.audit_attestation,
-    };
-    let bytes = serde_json::to_vec(&payload).map_err(|e| {
-        NonoError::Snapshot(format!("Failed to serialize session digest payload: {e}"))
-    })?;
-    let mut hasher = Sha256::new();
-    hasher.update(SESSION_DIGEST_DOMAIN);
-    hasher.update(bytes);
-    Ok(ContentHash::from_bytes(hasher.finalize().into()))
-}
-
-#[cfg(unix)]
-fn path_bytes(path: &std::path::Path) -> Vec<u8> {
-    path.as_os_str().as_bytes().to_vec()
-}
-
-#[cfg(not(unix))]
-fn path_bytes(path: &std::path::Path) -> Vec<u8> {
-    path.to_string_lossy().into_owned().into_bytes()
-}
 
 pub(crate) fn append_session(metadata: &SessionMetadata) -> Result<LedgerRecord> {
     validate_ledger_session_id(&metadata.session_id)?;
@@ -139,79 +39,7 @@ pub(crate) fn append_session(metadata: &SessionMetadata) -> Result<LedgerRecord>
                 path.display()
             ))
         })?;
-    append_locked(&mut file, metadata)
-}
-
-fn validate_ledger_session_id(session_id: &str) -> Result<()> {
-    let valid = !session_id.is_empty()
-        && session_id.len() <= 64
-        && session_id
-            .bytes()
-            .all(|b| b.is_ascii_alphanumeric() || matches!(b, b'-' | b'_'));
-    if valid {
-        Ok(())
-    } else {
-        Err(NonoError::ConfigParse(format!(
-            "invalid audit session id: {session_id}"
-        )))
-    }
-}
-
-fn append_locked(file: &mut std::fs::File, metadata: &SessionMetadata) -> Result<LedgerRecord> {
-    let mut contents = String::new();
-    file.seek(SeekFrom::Start(0))
-        .map_err(|e| NonoError::Snapshot(format!("Failed to seek audit ledger: {e}")))?;
-    file.read_to_string(&mut contents)
-        .map_err(|e| NonoError::Snapshot(format!("Failed to read audit ledger: {e}")))?;
-
-    let mut previous_chain = None;
-    let mut next_sequence = 0u64;
-    for (index, line) in contents.lines().enumerate() {
-        if line.trim().is_empty() {
-            continue;
-        }
-        let record: LedgerRecord = serde_json::from_str(line).map_err(|e| {
-            NonoError::Snapshot(format!(
-                "Failed to parse audit ledger line {}: {e}",
-                index.saturating_add(1)
-            ))
-        })?;
-        previous_chain = Some(record.chain_hash);
-        next_sequence = record.sequence.saturating_add(1);
-    }
-
-    let session_digest = compute_session_digest(metadata)?;
-    let completed_at = metadata
-        .ended
-        .clone()
-        .unwrap_or_else(|| metadata.started.clone());
-    let chain_hash = hash_ledger_link(
-        previous_chain.as_ref(),
-        next_sequence,
-        &metadata.session_id,
-        &session_digest,
-        &completed_at,
-    )?;
-    let record = LedgerRecord {
-        sequence: next_sequence,
-        prev_chain: previous_chain,
-        session_id: metadata.session_id.clone(),
-        session_digest,
-        completed_at,
-        chain_hash,
-    };
-
-    file.seek(SeekFrom::End(0))
-        .map_err(|e| NonoError::Snapshot(format!("Failed to seek audit ledger for append: {e}")))?;
-    let line = serde_json::to_vec(&record).map_err(|e| {
-        NonoError::Snapshot(format!("Failed to serialize audit ledger record: {e}"))
-    })?;
-    file.write_all(&line)
-        .and_then(|_| file.write_all(b"\n"))
-        .and_then(|_| file.sync_data())
-        .map_err(|e| NonoError::Snapshot(format!("Failed to append audit ledger record: {e}")))?;
-
-    Ok(record)
+    append_session_to_ledger_file(&mut file, metadata)
 }
 
 pub(crate) fn verify_session_in_ledger(
@@ -220,15 +48,7 @@ pub(crate) fn verify_session_in_ledger(
     let root = audit_root()?;
     let path = root.join(AUDIT_LEDGER_FILENAME);
     if !path.exists() {
-        return Ok(LedgerVerificationResult {
-            hash_algorithm: LEDGER_HASH_ALGORITHM.to_string(),
-            entry_count: 0,
-            session_digest: compute_session_digest(metadata)?,
-            session_found: false,
-            session_digest_matches: false,
-            ledger_chain_verified: false,
-            ledger_head: None,
-        });
+        return missing_ledger_verification_result(metadata);
     }
 
     let file = OpenOptions::new().read(true).open(&path).map_err(|e| {
@@ -237,76 +57,7 @@ pub(crate) fn verify_session_in_ledger(
             path.display()
         ))
     })?;
-    let reader = BufReader::new(file);
-    let expected_digest = compute_session_digest(metadata)?;
-
-    let mut previous_chain = None;
-    let mut entry_count = 0u64;
-    let mut ledger_head = None;
-    let mut session_found = false;
-    let mut session_digest_matches = false;
-
-    for (index, line) in reader.lines().enumerate() {
-        let line = line.map_err(|e| {
-            NonoError::Snapshot(format!(
-                "Failed to read audit ledger {}: {e}",
-                path.display()
-            ))
-        })?;
-        if line.trim().is_empty() {
-            continue;
-        }
-        let record: LedgerRecord = serde_json::from_str(&line).map_err(|e| {
-            NonoError::Snapshot(format!(
-                "Failed to parse audit ledger line {}: {e}",
-                index.saturating_add(1)
-            ))
-        })?;
-        if record.sequence != entry_count {
-            return Err(NonoError::Snapshot(format!(
-                "Audit ledger sequence mismatch at line {}",
-                index.saturating_add(1)
-            )));
-        }
-        if record.prev_chain != previous_chain {
-            return Err(NonoError::Snapshot(format!(
-                "Audit ledger prev_chain mismatch at line {}",
-                index.saturating_add(1)
-            )));
-        }
-        let chain_hash = hash_ledger_link(
-            previous_chain.as_ref(),
-            record.sequence,
-            &record.session_id,
-            &record.session_digest,
-            &record.completed_at,
-        )?;
-        if chain_hash != record.chain_hash {
-            return Err(NonoError::Snapshot(format!(
-                "Audit ledger chain hash mismatch at line {}",
-                index.saturating_add(1)
-            )));
-        }
-
-        if record.session_id == metadata.session_id {
-            session_found = true;
-            session_digest_matches = record.session_digest == expected_digest;
-        }
-
-        previous_chain = Some(record.chain_hash);
-        ledger_head = Some(record.chain_hash);
-        entry_count = entry_count.saturating_add(1);
-    }
-
-    Ok(LedgerVerificationResult {
-        hash_algorithm: LEDGER_HASH_ALGORITHM.to_string(),
-        entry_count,
-        session_digest: expected_digest,
-        session_found,
-        session_digest_matches,
-        ledger_chain_verified: true,
-        ledger_head,
-    })
+    verify_session_in_ledger_reader(BufReader::new(file), metadata)
 }
 
 struct LedgerLock {
@@ -337,42 +88,15 @@ impl LedgerLock {
     }
 }
 
-fn hash_ledger_link(
-    previous: Option<&ContentHash>,
-    sequence: u64,
-    session_id: &str,
-    session_digest: &ContentHash,
-    completed_at: &str,
-) -> Result<ContentHash> {
-    let payload = LedgerLinkPayload {
-        sequence,
-        session_id,
-        session_digest: *session_digest,
-        completed_at,
-    };
-    let payload_bytes = serde_json::to_vec(&payload).map_err(|e| {
-        NonoError::Snapshot(format!(
-            "Failed to serialize audit ledger link payload: {e}"
-        ))
-    })?;
-    let mut hasher = Sha256::new();
-    hasher.update(LEDGER_CHAIN_DOMAIN);
-    if let Some(prev) = previous {
-        hasher.update(prev.as_bytes());
-    } else {
-        hasher.update([0u8; 32]);
-    }
-    hasher.update(payload_bytes);
-    Ok(ContentHash::from_bytes(hasher.finalize().into()))
-}
-
 #[cfg(test)]
 #[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
     use crate::test_env::{ENV_LOCK, EnvVarGuard};
+    use nono::audit::compute_session_digest;
     use nono::undo::{
-        AuditAttestationSummary, ExecutableIdentity, NetworkAuditDecision, NetworkAuditMode,
+        AuditAttestationSummary, AuditIntegritySummary, ContentHash, ExecutableIdentity,
+        NetworkAuditDecision, NetworkAuditEvent, NetworkAuditMode,
     };
     #[cfg(unix)]
     use std::ffi::OsString;
